Simplify and generalise regex again
Turns out that the more specific regex did not catch all glosses — [run-away] was not removed, for whatever strange reason (all letters in the gloss were part of the regex).

The risk with the more general regex is that it might remove more than intended. But the risk is probably pretty low, so we’ll accept it.

Fixes #3.